% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/categorize_p_value.R
\name{categorize_p_value}
\alias{categorize_p_value}
\title{categorize whether a p-value belongs to the null distribution}
\usage{
categorize_p_value(p_value, null_cutoff = 0.005, uncertain_cutoff = 5e-04)
}
\arguments{
\item{p_value}{probability that sample fits within the null distribution}

\item{null_cutoff}{threshold at which to assign the sample as belonging to
the null distribution}

\item{uncertain_cutoff}{threshold at which to assign the sample as belonging to
the null distribution}
}
\value{
string of "reject" if the p-value belongs to the null distribution,
    "null" if the p-value belongs to the null distribution, and
    "uncertain" if it is unclear whether the p-value belongs to the null
    distribution or not.
}
\description{
categorize whether a p-value belongs to the null distribution
}
\examples{
categorize_p_value(0.000001)
categorize_p_value(0.1)
categorize_p_value(NA)
categorize_p_value(0.001, uncertain_cutoff=0.002)
}

