% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_convex_data.R
\name{chrom_num}
\alias{chrom_num}
\title{convert a chrom string to a number (make sex chroms 23 and beyond)}
\usage{
chrom_num(chrom)
}
\arguments{
\item{chrom}{chromosome ID as string eg  "1", "2" ... "X"}
}
\value{
chromosome as number, where X and Y are 23 and 24
}
\description{
convert a chrom string to a number (make sex chroms 23 and beyond)
}

