% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifer-package.r
\docType{package}
\name{CIFER}
\alias{CIFER}
\alias{CIFER-package}
\title{cifer: predicts inheritance of CNVs from exome read-depth.}
\description{
cifer: predicts inheritance of CNVs from exome read-depth.
}

