% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/call_exome_cnvs.R
\name{classify_exome_cnv}
\alias{classify_exome_cnv}
\title{function to classify single CNV based on sample IDs, and CNV coordinates}
\usage{
classify_exome_cnv(proband_id, maternal_id, paternal_id, chrom, start_pos,
  stop_pos, cnv = NA,
  DATAFREEZE_DIR = "/nfs/ddd0/Data/datafreeze/1133trios_20131218")
}
\arguments{
\item{proband_id}{sample ID for the proband (eg DDDP100001)}

\item{maternal_id}{sample ID for the proband's mother}

\item{paternal_id}{sample ID for the proband's father}

\item{chrom}{chromosome that the CNV is on (eg "1", "2", ..., "X")}

\item{start_pos}{start nucleotide of the CNV}

\item{stop_pos}{stop nucleotide of the CNV}

\item{cnv}{one row dataframe that contains the details of the CNV. Optional,
only used if we need to plot the CNV, when we need to note any
existing inheritance classification.}

\item{DATAFREEZE_DIR}{path to datafreeze directory containing files with family
relationships and sanger IDs}
}
\value{
inheritance classification as a string eg "paternal_inh", "de_novo" etc
}
\description{
function to classify single CNV based on sample IDs, and CNV coordinates
}

