% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inheritance_from_p_values.R
\name{classify_from_trio_p_values}
\alias{classify_from_trio_p_values}
\title{classify the inheritance state of the child's CNV call}
\usage{
classify_from_trio_p_values(mom_p_value, dad_p_value, proband_p_value)
}
\arguments{
\item{mom_p_value}{p-value for mom's scores belonging to the null model}

\item{dad_p_value}{p-value for dad's scores belonging to the null model}

\item{proband_p_value}{p-value for proband's scores belonging to the null model}
}
\value{
inheritance classification for the CNV eg("de_novo", "uncertain",
        "paternal_inh" etc)
}
\description{
classify the inheritance state of the child's CNV call
}
\examples{
mom_p = 0.1
dad_p = 0.1
child_p = 0.0001
classify_from_trio_p_values(mom_p, dad_p, child_p)
}

