% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/call_exome_cnvs.R
\name{classify_inheritance}
\alias{classify_inheritance}
\title{classify the inheritance state of the childs CNV call}
\usage{
classify_inheritance(mom_p_value, dad_p_value, proband_p_value)
}
\arguments{
\item{mom_p_value}{p-value for mom's scores belonging to the null model}

\item{dad_p_value}{p-value for dad's scores belonging to the null model}

\item{proband_p_value}{p-value for proband's scores belonging to the null model}
}
\value{
inheritance classification for the CNV eg("de_novo", "uncertain",
    "paternal_inh" etc)
}
\description{
classify the inheritance state of the childs CNV call
}

