% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_convex_data.R
\name{extract_lines_from_file}
\alias{extract_lines_from_file}
\title{quickly open a section of a file}
\usage{
extract_lines_from_file(path, file_pos, lines_n, chrom, start_pos, filesize,
  l2r_type)
}
\arguments{
\item{path}{path to L2R dataset}

\item{file_pos}{position in the file to jump ahead to}

\item{lines_n}{number of lines to read in}

\item{chrom}{chromosome string for current CNV}

\item{start_pos}{CNV start nucleotide position as integer}

\item{filesize}{size of L2R file in bytes}

\item{l2r_type}{string indicating the type of L2R scores to load, "adm3" or
"l2r"}
}
\value{
dataframe of probe information, with one row per probe
}
\description{
quickly open a section of a file
}

