% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_convex_data.R
\name{find_convex_files}
\alias{find_convex_files}
\title{find the convex data files for each DDD individual}
\usage{
find_convex_files(ddd, convex_dir)
}
\arguments{
\item{ddd}{data frame of participant informatyion, including sample IDs}

\item{convex_dir}{path to folder containing individual L2R datafiles}
}
\value{
dataframe with additional column of file paths
}
\description{
find the convex data files for each DDD individual
}

