% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_convex_data.R
\name{find_file_pos}
\alias{find_file_pos}
\title{find where the required data lies within the file}
\usage{
find_file_pos(path, file_pos, chrom, start_pos, file_size)
}
\arguments{
\item{path}{path to the datafile}

\item{file_pos}{reasonable guess about where the correct data is located in
the file}

\item{chrom}{chrom that we are searching for}

\item{start_pos}{start position of the CNV}

\item{file_size}{size of L2R file in bytes}
}
\value{
the approximate byte offset in the file (to within a few hundred
   bytes) where the required data starts
}
\description{
Seek in a file to the section containing the data that we want. We start
with a reasonable guess about where in the file the data is located, but
we seek up or downstream of that (depending on whether the data at that
position is before or after the data that we want), then bisect to pin
down where to start loading data from.
}

