% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/call_exome_cnvs.R
\name{get_ddd_individuals}
\alias{get_ddd_individuals}
\title{get a list of DDD participants, with their family relationships, and
sanger sample IDs}
\usage{
get_ddd_individuals(samples_dir)
}
\arguments{
\item{samples_dir}{path to datafreeze directory containing files with family
relationships and sanger IDs}
}
\value{
dataframe with sample information, including sanger IDs
}
\description{
get a list of DDD participants, with their family relationships, and
sanger sample IDs
}

