% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/call_exome_cnvs.R
\name{get_l2r_z_scores}
\alias{get_l2r_z_scores}
\title{get Z transformed L2R scores in the population, and in the proband family}
\usage{
get_l2r_z_scores(mother, father, proband, population)
}
\arguments{
\item{mother}{l2r values for the mother of the current proband, or NULL if
the mother is not present}

\item{father}{l2r values for the father of the current proband, or NULL if
the father is not present}

\item{proband}{l2r values for the current proband}

\item{population}{l2r values for parents not in proband's family}
}
\value{
list of z scores for population (as vector), and z scores for proband
    family (as list)
}
\description{
get Z transformed L2R scores in the population, and in the proband family
}
\examples{
get_l2r_z_scores(c(1,2), c(0,3), c(3,3), data.frame(a=c(0.5, 0.1), b=c(0.3, 0.2)))
}

