% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/determine_null_parameters.R
\name{get_maxima}
\alias{get_maxima}
\title{estimate local maxima within a numeric distribution}
\usage{
get_maxima(z_scores)
}
\arguments{
\item{z_scores}{numeric vector of Z-transformed log2-ratio data for parental
population}
}
\value{
vector of maxima positions
}
\description{
Note that this function tends to overestimate the number of maxima in the
distribution, which is compensated for by only picking the biggest of the
peaks later on.
}
\examples{
get_maxima(rnorm(100, mean=0, sd=1))
get_maxima(c(rnorm(100, mean=0, sd=1), rnorm(20, mean=4, sd=1)))
}

