% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/call_exome_cnvs.R
\name{get_null_parameters}
\alias{get_null_parameters}
\title{get the parameters of the null distrubution}
\usage{
get_null_parameters(z_scores)
}
\arguments{
\item{z_scores}{Z score transformed log2 ratio data for parents unrelated
to the proband currently being classified.}
}
\value{
a list containing the null model's mean and standard deviation, or
    raises an error if generating a mixture model and the code has too
    many retries
}
\description{
get the parameters of the null distrubution
}
\examples{
get_null_parameters(rnorm(100, mean=0, sd=1))
get_null_parameters(c(rnorm(100, mean=0, sd=1), rnorm(30, mean=5, sd=1)))
}

