% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{get_parental_correlations}
\alias{get_parental_correlations}
\title{find the correlation in L2R values between a child and their parents}
\usage{
get_parental_correlations(proband, mother, father)
}
\arguments{
\item{proband}{L2R values for proband (or NULL if not available)}

\item{mother}{L2R values for mother (or NULL if not available)}

\item{father}{L2R values for father (or NULL if not available)}
}
\value{
list of correlations for mother and father's L2R scores with probands
    L2R values, or NA where L2R values are unavailable.
}
\description{
find the correlation in L2R values between a child and their parents
}

