% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{get_population_correlation}
\alias{get_population_correlation}
\title{get the correlation between each proband and their parents L2R values}
\usage{
get_population_correlation(ddd, probes, proband_ids, parent_ids)
}
\arguments{
\item{ddd}{dataframe containing all the DDD sample information.}

\item{probes}{L2R or ADM3 scores for all individuals for probes within the CNV}

\item{proband_ids}{vector of proband IDs, except for the trio being investigated}

\item{parent_ids}{vector of parental IDs, except for the trio under
investigation, sorted as for the vector of parental Z scores.}
}
\value{
numeric vector of correlation scores, (with sample IDs as names),
    sorted according to the parent_ids vector.
}
\description{
get the correlation between each proband and their parents L2R values
}

