% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_convex_data.R
\name{get_probes_data}
\alias{get_probes_data}
\title{get probe scores for all the individuals, so that we can pick the values
for every individual at once}
\usage{
get_probes_data(ddd, chrom, start_pos, stop_pos, l2r_type = "adm3",
  convex_dir = "/lustre/scratch113/projects/ddd/users/ddd/DDD_ANALYSIS/convex/data",
  l2r_dir = "/lustre/scratch113/projects/ddd/users/jm33/aggregated_convex")
}
\arguments{
\item{ddd}{dataframe listing sample IDs, and file paths}

\item{chrom}{chromosome of interest as string (eg '1', '2', 'X')}

\item{start_pos}{chromosome position that the CNV region starts at}

\item{stop_pos}{chromosome position that the CNV region ends at}

\item{l2r_type}{type of log2-ratio data to load (either "l2r" or "adm3")}

\item{convex_dir}{path folder containing individual log2-ratio data}

\item{l2r_dir}{path to folder containing aggregated log2-ratio data}
}
\value{
a dataframe containing all the L2R data for the different probes in
    the CNV region, for all the particpants in the ddd dataframe
}
\description{
get probe scores for all the individuals, so that we can pick the values
for every individual at once
}

