% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{include_graphs}
\alias{include_graphs}
\title{set up the data for plotting a graph, then call the plotting function}
\usage{
include_graphs(ddd, probes, z_scores, proband_id, maternal_id, paternal_id, cnv)
}
\arguments{
\item{ddd}{dataframe containing all the DDD sample information.}

\item{probes}{L2R or ADM3 scores for all individuals for probes within the CNV}

\item{z_scores}{list of Z scores for unrelated parents, and Z scores for
members of the current trio}

\item{proband_id}{DDD ID of the proband}

\item{maternal_id}{DDD ID of the proband's mother}

\item{paternal_id}{DDD ID of the proband's father}

\item{cnv}{one row dataframe that contains the CNV's details (used to
identify the plot)}
}
\description{
Initially we expected that correlations between the proband's scores with
parental scores might be useful, which is why this function calculates the
correlation, before plotting.
}

