% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_convex_data.R
\name{open_probes_data_fast}
\alias{open_probes_data_fast}
\title{open convex data from an aggregated convex dataset}
\usage{
open_probes_data_fast(ddd, chrom, start_pos, stop_pos, l2r_type, l2r_dir)
}
\arguments{
\item{ddd}{dataframe listing sample IDs, and file paths}

\item{chrom}{chromosome of interest as string (eg '1', '2', 'X')}

\item{start_pos}{chromosome position that the CNV region starts at}

\item{stop_pos}{chromosome position that the CNV region ends at}

\item{l2r_type}{string indicating the type of L2R scores to load, "adm3" or
"l2r"}

\item{l2r_dir}{path to folder containing the aggregated L2R scores}
}
\value{
a dataframe containing all the data for the different probes in
    the CNV region, for all the particpants in the ddd dataframe
}
\description{
open convex data from an aggregated convex dataset
}

