% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_convex_data.R
\name{open_probes_data_slow}
\alias{open_probes_data_slow}
\title{parse all the convex files for the individuals, to get the L2R values
for every individual}
\usage{
open_probes_data_slow(ddd, chrom, start_pos, stop_pos, l2r_type)
}
\arguments{
\item{ddd}{dataframe listing sample IDs, and file paths}

\item{chrom}{chromosome of interest as string (eg '1', '2', 'X')}

\item{start_pos}{chromosome position that the CNV region starts at}

\item{stop_pos}{chromosome position that the CNV region ends at}

\item{l2r_type}{string indicating the type of L2R scores to load, "adm3" or
"l2r"}
}
\value{
a dataframe containing all the L2R data for the different probes in
    the CNV region, for all the particpants in the ddd dataframe
}
\description{
parse all the convex files for the individuals, to get the L2R values
for every individual
}

