% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_cluster}
\alias{plot_cluster}
\title{plot the Z scores and correlations for the population, so that we can also
plot the parental data points, in order to quickly assess how far away
from the normal population that parental data points are.}
\usage{
plot_cluster(z_scores, correlations, trio_corr, cnv)
}
\arguments{
\item{z_scores}{Z scores of L2R for the parental population}

\item{correlations}{correlations of parental L2R to child L2R for the
parental population}

\item{trio_corr}{correlations of mother's L2R to proband's L2R}

\item{cnv}{row of data frame for CNV containing the proband ID, chrom,
start, stop, manually reviewed call, which we use in the title of
the plot.}
}
\description{
plot the Z scores and correlations for the population, so that we can also
plot the parental data points, in order to quickly assess how far away
from the normal population that parental data points are.
}

