% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_cnv.R
\name{plot_cnv}
\alias{plot_cnv}
\title{plot the density of Z-scores from the population and trio}
\usage{
plot_cnv(z_scores, cnv)
}
\arguments{
\item{z_scores}{list of Z scores for unrelated parents, and Z scores for
members of the current trio}

\item{cnv}{one row dataframe that contains the CNV's details (used to
identify the plot)}
}
\description{
plot the density of Z-scores from the population and trio
}

