% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_graph}
\alias{plot_graph}
\title{set up the data for plotting a graph, then call the plotting function}
\usage{
plot_graph(z_scores, cnv)
}
\arguments{
\item{z_scores}{list of Z scores for unrelated parents, and Z scores for
members of the current trio}

\item{cnv}{one row dataframe that contains the CNV's details (used to
identify the plot)}
}
\description{
set up the data for plotting a graph, then call the plotting function
}

