% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/call_exome_cnvs.R
\name{predict_inheritance}
\alias{predict_inheritance}
\title{quantify the probability that the parental CNVs belong to the null
(uninherited) distribution}
\usage{
predict_inheritance(z_scores)
}
\arguments{
\item{z_scores}{list of Z scores of L2R for the parental population, and
Z scores for the mother, father and current proband}
}
\value{
list of inheritance classification, and P values for the mother and
    father, where the values indicate how likely each parents data belongs
    to the uninherited cluster.
}
\description{
quantify the probability that the parental CNVs belong to the null
(uninherited) distribution
}

