% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/z_transform_l2r_scores.R
\name{get_l2r_z_scores}
\alias{get_l2r_z_scores}
\title{get Z transformed L2R scores in the population, and in the proband family}
\usage{
get_l2r_z_scores(family, population)
}
\arguments{
\item{family}{list of l2r values for the current trio, ie values for mother,
father and child. The parent values should be NULL if the parents do not
have data available.}

\item{population}{l2r values for parents not in proband's family}
}
\value{
list of population Z scores (as vector), and family Z scores (as list)
}
\description{
get Z transformed L2R scores in the population, and in the proband family
}
\examples{
family = list("mom"=c(1,2), "dad"=c(0,3), "child"=c(3,3))
population = data.frame(a=c(0.5, 0.1), b=c(0.3, 0.2))
get_l2r_z_scores(family, population)
}

