% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_inheritance_from_z_scores.R
\name{predict_inheritance}
\alias{predict_inheritance}
\title{predict the inheritance of a child's CNV from population and family scores}
\usage{
predict_inheritance(population, family)
}
\arguments{
\item{population}{vector of Z scores of L2R for the parental population}

\item{family}{list of Z scores for "mom", "dad" and "child"}
}
\value{
list of inheritance classification, and P values for the mother and
    father, where the values indicate how likely each parents data belongs
    to the uninherited cluster.
}
\description{
predict the inheritance of a child's CNV from population and family scores
}
\examples{
family = list("mom"=0, "dad"=0, "child"=5)
population = rnorm(100)
predict_inheritance(population, family)

family = list("mom"=NA, "dad"=NA, "child"=1)
population = rnorm(100)
predict_inheritance(population, family)
}

