% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_cnv_call.R
\name{process_cnv_call}
\alias{process_cnv_call}
\title{predict the inheritance state of a CNV call}
\usage{
process_cnv_call(samples, probes, child_id, mom_id, dad_id, cnv = NA)
}
\arguments{
\item{samples}{dataframe listing sample IDs, and file paths}

\item{probes}{dataframe of log-2-ratio (or adm3 score) values for all
participants, for the exome probes that lie within the CNV region.}

\item{child_id}{ID of the proband}

\item{mom_id}{sample ID for the proband's mother}

\item{dad_id}{sample ID for the proband's father}

\item{cnv}{row information for CNV}
}
\value{
list of inheritance classification, the mother's p-value and the
       father's p-value.
}
\description{
predict the inheritance state of a CNV call
}
\examples{
samples = read.table(header=TRUE, text="
       individual_id is_proband
       A             FALSE
       B             FALSE
       C             FALSE
       D             FALSE
       E             TRUE
       F             FALSE
       G             FALSE
       H             FALSE
       I             FALSE
       J             FALSE
       K             TRUE
       L             FALSE",
       colClasses=c("character", "logical"))
probes = read.table(header=TRUE, text="
       probe A B C D E F G H I J K
       p1    5 6 4 5 5 6 5 5 4 6 9
       p2    5 6 4 5 5 6 5 5 4 6 9
       p3    5 6 4 5 5 6 5 5 4 6 9
       p4    5 6 4 5 5 6 5 5 4 6 9")
child_id = "K"
mom_id = "J"
dad_id = "I"
process_cnv_call(samples, probes, child_id, mom_id, dad_id)

# run another example, this time with a larger dataset
cohort_n = 500
sample_ids = paste("sample", 1:cohort_n, sep="_")
samples = data.frame("individual_id"=sample_ids,
    "is_proband"=c(rep(FALSE, length(sample_ids) - 1), TRUE))

# define the population as having probes values cenetred around zero
probes = data.frame(matrix(rnorm(length(sample_ids) * 5), nrow=5))
names(probes) = sample_ids

# set the child probe values to distant from the population values
probes[sample_ids[length(sample_ids)]] = rnorm(5, mean=10, sd=1)

# define the sample IDs for the trio members
child_id = sample_ids[length(sample_ids)]
mom_id = sample_ids[1]
dad_id = sample_ids[2]

process_cnv_call(samples, probes, child_id, mom_id, dad_id)
}

