% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/call_exome_cnvs.R
\name{process_cnv_call}
\alias{process_cnv_call}
\title{predict the inheritance state of the CNV call}
\usage{
process_cnv_call(ddd, probes, proband_id, maternal_id, paternal_id, cnv = NA)
}
\arguments{
\item{ddd}{dataframe listing sample IDs, and file paths}

\item{probes}{dataframe of log-2-ratio (or adm3 score) values for all DDD
participants, for the exome probes that lie within the CNV region.}

\item{proband_id}{ID of the proband}

\item{maternal_id}{DDD ID for the proband's mother}

\item{paternal_id}{DDD ID for the proband's father}

\item{cnv}{row information for CNV}
}
\value{
list of inheritance classification, the mother's p-value and the
    father's p-value.
}
\description{
predict the inheritance state of the CNV call
}

